+++ src/main/java/FinancialSystem/trans/SalaryInfo.java
@@ -4,12 +4,12 @@ public class SalaryInfo {
     private String name;
     private String workerid;
     private String banknum;
-    private int money;
+    private double money;
     private String type;
     public SalaryInfo(){
 
     }
-    public SalaryInfo(String name,String workerid,String banknum,int money,String type){
+    public SalaryInfo(String name,String workerid,String banknum,double money,String type){
         this.name = name;
         this.workerid = workerid;
         this.banknum = banknum;
@@ -42,11 +42,11 @@ public class SalaryInfo {
         this.banknum = banknum;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
 
-    public void setMoney(int money) {
+    public void setMoney(double money) {
         this.money = money;
     }
